Testing

+++ King.Azure.Imaging.Mvc/Controllers/api/SecureImageController.cs
@@ -35,7 +35,7 @@
 
         [HttpGet]
         [Authorize(Roles = "Users")]
-        public override async Task<HttpResponseMessage> Get(string file, int width = 0, int height = 0, string format = Naming.DefaultExtension, int quality = Imaging.DefaultImageQuality, bool cache = true)
+        public override async Task<HttpResponseMessage> Get(string file, ushort width = 0, ushort height = 0, string format = Naming.DefaultExtension, byte quality = Imaging.DefaultImageQuality, bool cache = true)
         {
             return await base.Get(file, width, height, format, quality, cache);
         }
